Ejercicio1/Functions/requirements.txt
gcsfs==2022.8.2
fsspec==2022.8.2
google-cloud-storage==2.5.0
pandas==1.3.5
pandasql==0.7.3
numpy==1.21.6